requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
asyncpg>=0.29.0
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
locust>=2.17.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0
